Extract shared state + marker-file writers from process_control.py
Second (foundational + biggest cluster) extraction from
modules/process_control.py.  Sets up the bottom-layer state module
that later extractions will build on.

## What moved

### modules/process_control_state.py (~93 lines, new)

Bottom-layer module that owns:

* Runtime-state dicts: MAINTENANCE_STATE, PENDING_KOMETA_START,
  RUN_CONTEXT, IMAGEMAID_RUN_CONTEXT
* Their threading.Lock companions
* MAINTENANCE_GUARD_INTERVAL constant
* normalize_kometa_start_mode (a 2-line pure function called from
  4 different clusters)

Nothing imports from any sibling -- this file is the foundation the
other extractions will import from.  Breaks the circular-import
problem cleanly: process_control.py re-exports from process_markers,
process_markers imports from process_control_state, and neither
imports from the other.

### modules/process_markers.py (~380 lines, new)

19 marker-file writer functions covering Kometa and ImageMaid run
markers, maintenance sidecars, meta.log appenders, and config-path
resolution.

## modules/process_control.py: 1155 -> 839 (-316 / -27.4%)

Cleaned up newly-unused imports (threading, current_app,
has_app_context) since their only remaining users moved out.

## Backward compatibility

process_control.py re-exports every moved name via three re-export
blocks (process_metrics + process_control_state + process_markers).
All 43 imports in quickstart.py (using X as _X aliases) continue
to work unchanged.

## Verification

AST-verified: all 20 moved functions are byte-for-byte identical
to origin/develop.  Full test suite passes (0 failures).
